pyproject.toml
[project]
name = "PersonaEval"
version = "0.1.0"
description = "A comprehensive benchmark for LLM judges of role-play"
authors = [
    {name = "Lingfeng Zhou", email = "zhoulingfeng@sjtu.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "pandas>=2.0.0",
    "requests[socks]>=2.31.0",
    "openai>=1.0.0",
    "tqdm>=4.65.0",
    "json-repair>=0.19.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.0.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
personaeval = "personaeval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 